Move author directory to public folder and add to version control
The author dir doesn't need to be in app directory for the app to be
compiled correctly and will just slow down the build if the author
vidoes needs to be copied over each time. Therefore we move it to
public permanently and add the empty folder to version control to make
it clear that's where the videoes needs to be placed for the app to work
correctly.